modified philosopher. I had to use while loop to wait precisely.

+++ lv3/philosophers/philo/utils/time_ms.c
@@ -6,13 +6,13 @@
 /*   By: hmiyazak <hmiyazak@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/12 14:47:44 by hmiyazak          #+#    #+#             */
-/*   Updated: 2023/11/25 20:10:35 by hmiyazak         ###   ########.fr       */
+/*   Updated: 2024/04/28 21:15:47 by hmiyazak         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../philo.h"
 
-int	get_time_ms(double *time)
+int	get_time_ms(long int *time)
 {
 	t_time	current_time;
 
@@ -20,11 +20,11 @@ int	get_time_ms(double *time)
 		return (-1);
 	if (gettimeofday(&current_time, NULL) != 0)
 		return (-1);
-	*time = current_time.tv_sec * 1000.0 + current_time.tv_usec / 1000.0;
+	*time = current_time.tv_sec * 1000 + current_time.tv_usec / 1000;
 	return (0);
 }
 
-double	create_timestamp(double c_time, t_table *table)
+long int	create_timestamp(long int c_time, t_table *table)
 {
 	return (c_time - table->start_time);
 }
